# 校园二手交易APP/openapi.yaml
openapi: 3.0.3
info:
  title: 校园二手交易 APP API
  version: 1.0.0
  description: |
    Android 客户端使用的校园二手交易平台后端接口。所有定义与 `api.md`、`db_schema.md.md`、`spec.md`、
    《后端 API 接口开发规划文档》保持一致。
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.campus-trade.com/api/v1
    description: 生产示例
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: OK
        errorCode:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        data:
          description: 具体数据载荷
          nullable: true
    PaginatedMeta:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
    UserSummary:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role:
          type: string
          enum: [STUDENT, TEACHER, STAFF]
        email: { type: string }
        phone: { type: string, nullable: true }
        avatar: { type: string, nullable: true }
        school: { type: string, nullable: true }
        emailVerified: { type: boolean }
    Product:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        price: { type: number, format: double }
        category:
          type: string
          enum: [BOOKS, ELECTRONICS, CLOTHING, SPORTS, DAILY, OTHER]
        categoryId:
          type: integer
          nullable: true
          description: 动态分类ID（可选）
        categoryCode:
          type: string
          nullable: true
          description: 动态分类编码（可选）
        categoryName:
          type: string
          nullable: true
          description: 动态分类名称（可选）
        images:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [DRAFT, ON_SALE, OFF_SALE, SOLD, DELETED]
        auditStatus:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        viewCount: { type: integer }
        likeCount: { type: integer }
        contactInfo: { type: string, nullable: true }
        location: { type: string, nullable: true }
        remark: { type: string, nullable: true }
        seller:
          $ref: '#/components/schemas/UserSummary'
        createTime: { type: string, format: date-time }
        favorited: { type: boolean }
...
    AddressRequest:
      type: object
      required: [receiverName, phone, province, city, detailAddress]
      properties:
        receiverName: { type: string, maxLength: 50 }
        phone: { type: string, maxLength: 20 }
        province: { type: string, maxLength: 50 }
        city: { type: string, maxLength: 50 }
        district: { type: string, maxLength: 50, nullable: true }
        detailAddress: { type: string, maxLength: 200 }
        isDefault: { type: boolean, nullable: true }
    AddressResponse:
      type: object
      properties:
        id: { type: integer }
        receiverName: { type: string }
        phone: { type: string }
        province: { type: string }
        city: { type: string }
        district: { type: string, nullable: true }
        detailAddress: { type: string }
        isDefault: { type: boolean }
...
    ReportCreateRequest:
      type: object
      required: [targetType, reason]
      properties:
        targetType:
          type: string
          enum: [PRODUCT, USER, ORDER, MESSAGE, OTHER]
        targetId:
          type: integer
          nullable: true
        reason:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          nullable: true
        evidenceUrls:
          type: array
          maxItems: 5
          items: { type: string }
        contactInfo:
          type: string
          maxLength: 100
          nullable: true
    ReportResponse:
      type: object
      properties:
        id: { type: integer }
        targetType:
          type: string
          enum: [PRODUCT, USER, ORDER, MESSAGE, OTHER]
        targetId: { type: integer, nullable: true }
        targetSnapshot: { type: string, nullable: true }
        reason: { type: string }
        description: { type: string, nullable: true }
        evidenceUrls:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, RESOLVED, REJECTED]
        contactInfo: { type: string, nullable: true }
        resolution: { type: string, nullable: true }
        handledBy: { type: string, nullable: true }
        handledTime: { type: string, format: date-time, nullable: true }
        autoFlagged: { type: boolean }
        autoReason: { type: string, nullable: true }
        reporter:
          $ref: '#/components/schemas/UserSummary'
        createTime: { type: string, format: date-time }
    ReportDetailResponse:
      type: object
      properties:
        id: { type: integer }
        targetType:
          type: string
          enum: [PRODUCT, USER, ORDER, MESSAGE, OTHER]
        targetId: { type: integer, nullable: true }
        targetSnapshot: { type: string, nullable: true }
        reason: { type: string }
        description: { type: string, nullable: true }
        evidenceUrls:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, RESOLVED, REJECTED]
        contactInfo: { type: string, nullable: true }
        resolution: { type: string, nullable: true }
        handledBy: { type: string, nullable: true }
        handledTime: { type: string, format: date-time, nullable: true }
        autoFlagged: { type: boolean }
        autoReason: { type: string, nullable: true }
        createTime: { type: string, format: date-time }
        reporter:
          $ref: '#/components/schemas/UserSummary'
    ReportAuditRequest:
      type: object
      required: [status, resolution]
      properties:
        status:
          type: string
          enum: [RESOLVED, REJECTED]
        resolution:
          type: string
          maxLength: 200
...
paths:
...
  /users/me/addresses:
    get:
      summary: 获取我的地址列表
      responses:
        '200':
          description: 地址列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AddressResponse'
    post:
      summary: 新增地址
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AddressResponse'
  /users/me/addresses/{id}:
    put:
      summary: 更新地址
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AddressResponse'
    delete:
      summary: 删除地址
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /users/me/addresses/{id}/default:
    post:
      summary: 设为默认地址
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AddressResponse'
...
  /products:
    get:
      summary: 商品列表
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 10, maximum: 50 }
        - in: query
          name: category
          schema: { type: string, enum: [BOOKS, ELECTRONICS, CLOTHING, SPORTS, DAILY, OTHER] }
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: sortBy
          schema: { type: string, enum: [createTime, price, viewCount], default: createTime }
        - in: query
          name: sortOrder
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/Product' }
                          meta:
                            $ref: '#/components/schemas/PaginatedMeta'
    post:
      summary: 发布商品
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, price, category]
              properties:
                title: { type: string }
                description: { type: string }
                price: { type: number }
                category: { type: string, enum: [BOOKS, ELECTRONICS, CLOTHING, SPORTS, DAILY, OTHER] }
                images:
                  type: array
                  items: { type: string }
                contactInfo: { type: string }
                location: { type: string }
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
...
  /reports:
    post:
      summary: 提交举报
      description: 举报商品、用户、订单或聊天消息，支持附带最多 5 张证据截图
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '200':
          description: 举报已接收
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      summary: 我的举报
      description: 按状态/类型筛选当前用户提交的举报记录
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, IN_PROGRESS, RESOLVED, REJECTED]
        - in: query
          name: targetType
          schema:
            type: string
            enum: [PRODUCT, USER, ORDER, MESSAGE, OTHER]
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 10, maximum: 50 }
      responses:
        '200':
          description: 举报列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/ReportResponse'
                          meta:
                            $ref: '#/components/schemas/PaginatedMeta'
  /reports/{id}:
    get:
      summary: 举报详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 举报详细信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportDetailResponse'
...
  /profile:
    get:
      summary: 个人资料
      responses:
        '200':
          description: 资料
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserSummary'
    put:
      summary: 更新资料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
                school: { type: string }
                avatar: { type: string }
                contactInfo: { type: string }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /profile/products:
    get:
      summary: 我的商品（个人中心入口）
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 10 }
      responses:
        '200': { description: 列表, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
...
  /profile/delete-request:
    post:
      summary: 申请注销
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
...
  /admin/moderation/reports/{id}:
    get:
      summary: 举报详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 举报详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReportDetailResponse'
    patch:
      summary: 处理举报
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportAuditRequest'
      responses:
        '200':
          description: 举报已处理
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...